Feat: add searchHandler fn on SearchBox for keydown and click
