+++ client/src/store/useSearchStore.js
@@ -1,0 +1,16 @@
+import { create } from "zustand";
+
+const useSearchStore = create((set) => ({
+  inputValue: "",
+  fetchedRestaurants: [],
+  searchLoading: false,
+  searchError: false,
+
+  setInputValue: (inputValue) => set({ inputValue }),
+  setFetchedRestaurants: (restaurants) =>
+    set({ fetchedRestaurants: restaurants }),
+  setSearchLoading: (status) => set({ searchLoading: status }),
+  setSearchError: (status) => set({ searchError: status }),
+}));
+
+export default useSearchStore;
